+++ app/components/buttons_component.html.erb
@@ -1,6 +1,6 @@
 <div class="row justify-content-end">
   <div class="col-auto">
-    <%= form.submit 'Save as draft', class: 'btn btn-primary' %>
+    <%= form.submit 'Save as draft', class: 'btn btn-primary', id: 'save-draft-button' %>
     <%= form.submit 'Deposit', class: 'btn btn-primary' %>
   </div>
 </div>
